prep ztf join
