+++ .github/workflows/run_test.yml
@@ -178,6 +178,11 @@ jobs:
         source scripts/start_services.sh --kafka-version ${KAFKA_VERSION} --hbase-version ${HBASE_VERSION}
         cd $FINK_GRB
 
+    - name: Generate fake gcn counterparts
+      run: |
+        python fink_grb/test/prep_ztf_data.py
+        ls -lth fink_grb/test/test_data/ztf_test/online/raw/year=2019/month=09/day=03/
+
     - name: Call raw2science
       run: |
         fink start raw2science -c ${FINK_GRB}/fink_grb/test/test_data/with_hbase/fink_test.conf --night $NIGHT --exit_after 40
