+++ fink_grb/conftest.py
@@ -8,7 +8,6 @@ from astropy.time import Time
 
 from fink_grb.gcn_stream.gcn_reader import load_voevent_from_path
 from fink_grb.observatory import voevent_to_class
-from fink_grb.utils.fun_utils import get_observatory
 
 
 @pytest.fixture(autouse=True)
@@ -228,26 +227,3 @@ def init_spark(doctest_namespace):
     if np.__version__ >= "1.14.0":
         np.set_printoptions(legacy="1.13")
 
-
-def spatial_time_align(ztf_raw_data, grb_pdf):
-    ztf_raw_data = ztf_raw_data.copy()
-
-    all_obs = grb_pdf["raw_event"].map(get_observatory).values
-    random_obs = np.random.choice(all_obs, int((len(all_obs) + 1) / 2))
-    jd_gcn = [obs.get_trigger_time()[1] for obs in random_obs]
-
-    rand_ztf_index = np.random.choice(ztf_raw_data.index, len(random_obs))
-    for rows_ztf_cand, rows_ztf_prv, new_jd in zip(
-        ztf_raw_data.loc[rand_ztf_index, "candidate"],
-        ztf_raw_data.loc[rand_ztf_index, "prv_candidates"],
-        jd_gcn,
-    ):
-        rows_ztf_cand["jdstarthist"] = new_jd + np.random.uniform(1, 5)
-        rows_ztf_prv[0]["jd"] = new_jd + np.random.uniform(0.1, 1)
-
-    rand_ztf_index = np.random.choice(ztf_raw_data.index, len(random_obs))
-    for rows_ztf, new_jd in zip(ztf_raw_data.loc[rand_ztf_index, "candidate"], jd_gcn):
-        rows_ztf["jdstarthist"] = new_jd + np.random.uniform(1, 5)
-        rows_ztf["prv_candidates"] = None
-
-    return ztf_raw_data
